refactor: kopisAPI, concertAPI 분리
- kopisAPI 파일에는 kopis 오픈 API 관련 로직, concertAPI 파일에는 firebase 관련 로직이 들어가도록 분리
- 함수 선언을 화표 함수로 변경 (팀 컨벤션)
